Close stdin on Stream completion
When the Stream passed to Process.stdin is finalized, we close the
stdin channel we had with the spawned process. Meaning it can not be
reused in the future, if you are in need of pausing writes to the stdin
please do so without terminating the stream.
I might make this an option, but if I do we need to expose another way of
closing the stdin channel. For now this seems the most natural way of doing
so.